# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VERGA API",
        "description": "API REST pour les applications externes VERGA (back-office agence Angular, application client mobile/web). Authentification Bearer Sanctum.",
        "contact": {
            "name": "VERGA",
            "email": "contact@verga.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
        "/agence/register": {
            "post": {
                "tags": [
                    "Agence - Auth"
                ],
                "summary": "Inscription agence",
                "description": "Crée un compte gérant (`role=agence`) et le profil agence associé. Retourne un token Bearer pour connexion immédiate.",
                "operationId": "agenceRegister",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "email",
                                    "telephone",
                                    "gerant_name",
                                    "gerant_email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Transit Express Libreville"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "contact@transit-express.test"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "0612345678"
                                    },
                                    "type_agence_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    },
                                    "ville": {
                                        "type": "string",
                                        "example": "Libreville",
                                        "nullable": true
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "pays": {
                                        "type": "string",
                                        "example": "Gabon",
                                        "nullable": true
                                    },
                                    "gerant_name": {
                                        "type": "string",
                                        "example": "Jean Mbaye"
                                    },
                                    "gerant_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "gerant@transit-express.test"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "angular-backoffice"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé + token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/TokenResponse"
                                        },
                                        {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "name": {
                                                            "type": "string"
                                                        },
                                                        "email": {
                                                            "type": "string"
                                                        },
                                                        "role": {
                                                            "type": "string",
                                                            "example": "agence"
                                                        },
                                                        "agence": {
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation échouée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Trop de tentatives"
                    }
                }
            }
        },
        "/agence/login": {
            "post": {
                "tags": [
                    "Agence - Auth"
                ],
                "summary": "Connexion agence",
                "description": "Authentifie le gérant de l'agence (email du compte `users`) et retourne un token Bearer.",
                "operationId": "agenceLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "gerant@agence.test"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "angular-backoffice"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/TokenResponse"
                                        },
                                        {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "name": {
                                                            "type": "string"
                                                        },
                                                        "email": {
                                                            "type": "string"
                                                        },
                                                        "role": {
                                                            "type": "string",
                                                            "example": "agence"
                                                        },
                                                        "agence": {
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Agence bloquée ou suspendue",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Identifiants invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Trop de tentatives"
                    }
                }
            }
        },
        "/agence/logout": {
            "post": {
                "tags": [
                    "Agence - Auth"
                ],
                "summary": "Déconnexion agence",
                "operationId": "agenceLogout",
                "responses": {
                    "200": {
                        "description": "Token révoqué",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/agence/me": {
            "get": {
                "tags": [
                    "Agence - Auth"
                ],
                "summary": "Profil agence connectée",
                "operationId": "agenceMe",
                "responses": {
                    "200": {
                        "description": "Profil utilisateur et agence"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Compte non agence ou agence inactive"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "Identifiants incorrects."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "TokenResponse": {
                "properties": {
                    "token": {
                        "type": "string",
                        "example": "1|abcdef..."
                    },
                    "token_type": {
                        "type": "string",
                        "example": "Bearer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token Sanctum obtenu via login ou register",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}